+++ map/autoware_map_loader/CMakeLists.txt
@@ -34,7 +34,7 @@ ament_auto_add_library(lanelet2_map_loader_node SHARED
   src/lanelet2_map_loader/lanelet2_map_loader_node.cpp
   src/lanelet2_map_loader/utils/lanelet2_map_loader_utils.cpp
   src/lanelet2_map_loader/utils/lanelet2_map_cell_metadata.cpp
-  src/lanelet2_map_loader/utils/lanelet2_selected_map_loader_module.cpp
+  src/lanelet2_map_loader/lanelet2_selected_map_loader_module.cpp
 )
 
 autoware_agnocast_wrapper_register_node(lanelet2_map_loader_node
