+++ map/autoware_map_loader/src/lanelet2_map_loader/lanelet2_selected_map_loader_module.hpp
@@ -12,10 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef LANELET2_MAP_LOADER__UTILS__LANELET2_SELECTED_MAP_LOADER_MODULE_HPP_
-#define LANELET2_MAP_LOADER__UTILS__LANELET2_SELECTED_MAP_LOADER_MODULE_HPP_
+#ifndef LANELET2_MAP_LOADER__LANELET2_SELECTED_MAP_LOADER_MODULE_HPP_
+#define LANELET2_MAP_LOADER__LANELET2_SELECTED_MAP_LOADER_MODULE_HPP_
 
-#include "lanelet2_map_cell_metadata.hpp"
+#include "utils/lanelet2_map_cell_metadata.hpp"
 
 #include <autoware/agnocast_wrapper/autoware_agnocast_wrapper.hpp>
 #include <autoware/agnocast_wrapper/node.hpp>
@@ -66,4 +66,4 @@ private:
 
 }  // namespace autoware::map_loader
 
-#endif  // LANELET2_MAP_LOADER__UTILS__LANELET2_SELECTED_MAP_LOADER_MODULE_HPP_
+#endif  // LANELET2_MAP_LOADER__LANELET2_SELECTED_MAP_LOADER_MODULE_HPP_
