+++ map/autoware_map_loader/test/lanelet2_map_loader/test_lanelet2_selected_map_loader_module.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "../src/lanelet2_map_loader/utils/lanelet2_selected_map_loader_module.hpp"
+#include "../src/lanelet2_map_loader/lanelet2_selected_map_loader_module.hpp"
 
 #include <ament_index_cpp/get_package_share_directory.hpp>
 #include <autoware/agnocast_wrapper/autoware_agnocast_wrapper.hpp>
